+++ FrenRider/Models/CharacterConfig.cs
@@ -75,8 +75,8 @@ public class CharacterConfig
     public string AutoRotationType { get; set; } = "FRENRIDER";
     public string AutoRotationTypeDD { get; set; } = "DD";
     public string AutoRotationTypeFATE { get; set; } = "FATE";
-    public int RotationPlugin { get; set; } = 2; // 0=BMR, 1=VBM, 2=RSR, 3=WRATH
-    public int RotationPluginForay { get; set; } = 3; // 0=BMR, 1=VBM, 2=RSR, 3=WRATH
+    public int RotationPlugin { get; set; } = 2; // 0=BMR, 1=VBM, 2=RSR, 3=WRATH, 4=DAEDALUS
+    public int RotationPluginForay { get; set; } = 3; // 0=BMR, 1=VBM, 2=RSR, 3=WRATH, 4=DAEDALUS
     public bool ForceBossModPresetRegardlessOfRotation { get; set; } = false;
     public int BossModAI { get; set; } = 0; // 0=on, 1=off
     public int PositionalInCombat { get; set; } = 3; // 0=Front, 1=Rear, 2=Any, 3=Auto